requirements.txt
django-rest-framework==0.1.0
Django==1.11
mysqlclient==1.3.12
markdown==3.0.1
django-filter==2.0.0
djangorestframework-jwt==1.11.0